Fix for another bug revealed by validation : flow capacity leak.
The balancer did not free an index for a flow to an expired backend.
This might lead to a premature flow table exhaustion.

The bug is not visible on the TP verification step, because it is not
in any datastructure in particular.

The bug is not visible during symbex, because symbex is too low level to
capture the semantic relations.

The bug appeared as a call trace that failed to validate.