Strenghten the nat secucity: drop spoofing packets.
Here a mismatch between the implementation and the spec has led to more secure
spec and implementation.
In particular, for external packets, implementation changed only the necessary
header fields before forwarding, while the spec assumed all the tcp/ip routing
related fields would be set according to the flow in the flow table.
Reconciliation resulted in the additional check, that ensures that the coming
packet belongs to the flow found in the table.