Trace the chunks with the fixed length correctly.
Trace the chunks with variable length only by the first byte. Currently the only
variable-length chunk for the 3 NFs we have is the ip-options section, which we
care about only as long as we can reach the TCP header. Later will need a finer
tracing for var-length sections, as they will play a key role for a DNS cache.